requirements.txt
Django==5.2.5
djangorestframework==3.15.2
mysqlclient==2.2.6
redis==5.2.1
celery==5.4.0
django-cors-headers==4.6.0
python-dotenv==1.0.1
gunicorn==23.0.0
django-extensions==3.2.3
google-generativeai==0.3.2